requirements.txt
Flask
mysql-connector-python
pandas
openpyxl
fpdf
Flask-Mail
WeasyPrint
python-dotenv
itsdangerous
werkzeug
